Extract libinput stuff
